fix: handle time changes via steady_clock (#51)
This resolves an issue where the timer thread could get _reall_ confused
if a time change (e.g. jumping from 1970 to current time) happens at
the wrong time of the timer loop. This could lead to the timer thread
basically hanging on a for loop iterating for every millisecond of
delta between the before (1970) and after (2021) times.

The culprit was the use of std::chrono::high_resolution_clock which
I mistakenly believed to be monotonic time. It's actually wall clock
time. This is trivally resolved by switching to std::chrono::steady_clock
instead which is guaranteed monotonic and immune to this time change
vulnerability in the logic.