fix: reduce timer stuttering by merging the timer and run threads (single threaded) (#92)
We've observed quite a bit of timer stuttering on some architectures in
the field. The goal of this change is to reduce that stuttering by
merging the timer and run threads in the `SingleThreadScheduler`.

The theory here is that the stuttering is partially induced by the
additional thread hop required for submitting a "ready" timer task to
the ready queue and then waking the run thread to process that work.
This would likely be especially bad in cases where the run thread spends
a lot of time idling - which can lead to varying times required for the
kernel to wake the run thread and do additional work.

Instead, this change makes the run thread handle both situations. It'll
first wake up and evaluate if any timers need to be executed. It'll then
run a batch of tasks from the ready queue. Finally it'll go idle, just
like before, but instead it'll schedule itself to wake up automatically
when the next timer is supposed to fire.

To simplify some of the code I've also added the `SpinLock` class which
is just a simple atomic spin lock wrapper. I've also spent some time
documenting when the relaxed memory ordering was used to justify that
usage.