BREAKING: Migrate `ThreadPoolScheduler` to `WorkStealingScheduler` (#93)
This change aims to improve the performance of cask while also
simplifying its implementation. In a production application we've
noticed some issues with performance stuttering that led to #92 and is
now leading to similar fixes in the multi-threaded scheduler. Rather
than do that work _twice_ this was a good opportunity to dust off the
the `ThreadPoolScheduler` and, instead, make it operate as a pool of
`SingleThreadScheduler` instances.

That design lends itself nicely to another big improvement - [Work
Stealing](https://en.wikipedia.org/wiki/Work_stealing).

This is a well-known and widely implemented strategy for implementing
multi-threaded schedulers that has significant benefits over the
approach of having a big queue that bunch of threads consume from.
Instead of having one big ready queue that all threads in a pool consume
from:

- Each thread has its own ready queue.
- When work is submitted back to the ready queue - the ready queue for
the _currently active_ thread is chosen if available. Otherwise a random
scheduler is chosen.
- When a thread runs _out_ of work it starts looking to the ready queue
of _another_ thread to "steal" work from it.

The result is that we never contend for one "big fat lock" a global (to
the pool) ready queue. That can really get out of control in terms of
contention especially as the number of threads grows. In addition - we
end up preferring to _continue_ work on the same thread that has already
been running which improves practical concerns like cache locality.

Finally, it means I only have to implement a bunch of scheduler things
_once_. I think you'll see that the logic the `WorkStealingScheduler`
adds to coordinate a pool of `SingleThreadScheduler` instances is very
straightforward. This means we get to share a lot of code between those
two things - whereas `ThreadPoolScheduler` was a totally independent
thing.

In order to increase confidence I've parameterized many of the test
benches so that they run across multiple scheduler types. I've also
converged the multiple scheduler test beds into a single parametrized
one.

This is a breaking change both due to compile time changes (I've removed
some things, changed some parameters, etc) and runtime changes (whole
new scheduler bro) - but in terms of usability it's a minor one. If
someone was using `ThreadPoolScheduler` directly before they'll need to
change there include/usings/etc but that's about it. If
`Scheduler::global` was used then no changes should be necessary to get
things compiling and running with the new version.