fix: resolve data races when racing fibers (#94)
This change resolves several independent issues which could all lead to
various data races when racing fibers that were either (a) delayed or
(b) waiting on some async boundary. Both of these cases were pretty
common and could lead to various crashes I've seen running this code in
the field.

To reveal these issues a single test was added that specifically
exercises this code path as a stress test. This test case revealed
several issues which all contributed to potentially buggy behavior with
anything that performed a `race` (like `timeout`) or tried to call
`cancel` from code executing concurrently to the thread a fiber was
executing on.

To get that stress test to run cleanly, the following issues were
resolved:

1. There was a small chance than and async or delay callback could race
with a concurrent cancel. In these situations the callback would observe
a fiber in the `RUNNING` state (because the cancel was being processed)
rather than the expected `WAITING` or `DELAYED` state. The resolution is
to busy wait inside the callback for the fiber to stop `RUNNING`
indicating that the cancel call has finished its processing - and then
continue processing the callback.
2. There was a chance that a delay could be de-allocated _while_ it was
processing a cancel callback - leading to weird mutex errors due to
corrupted data. The resolution here is applied in several different
ways:

- The timer logic has been refactored so that `SingleThreadScheduler`
holds onto instances of `CancelableTimer` until it is done with them
(either by firing the timer or seeing it canceled). This means that an
external observer simply dropping the handle won't invalidate these.
- `FiberImpl` copies the `delayedBy` and `waitingOn` references to local
copies before releasing atomic state and calling `cancel` to avoid a
race where a parallel thread completes a callback and nulls those
pointers.
- Callbacks are `std::swap`ed into a stack-local store before executing
them, to further prevent issues with these callbacks turning around and
messing with shared state leading to crashes.

3. When racing fibers, there was previously a chance that multiple
racing calls to `racerFinished` could _both_ observe that the queue was
empty resulting in multiple "completions" of the monitor fiber in
concurrent contexts - leading to data corruption. The logic of
`racerFinished` has been refactored to avoid this issue by making all
important decisions inside of a single lock hold rather than releasing a
lock, messing with atomics, and then locking again - which was far
harder to reason about.

Finally, to avoid these issues with callbacks not firing correctly in
the future logic was added to throw a `std::runtime_error` if some state
is encountered in those callbacks that isn't valid - which should lead
to finding these issues quicker in the future.