fix: avoid queue put and takes lists filled with canceled tasks (#95)
We've seen a crash where sometimes queue takes and puts can get into a
bad access pattern where the those lists end up filling with _canceled_
promises that don't get reaped because they are _deep_ in the list
_after_ some valid entry and `dropWhile` doesn't catch them.

The end result is that _freeing_ a `QueueState` can end up causing a
free of a large `List` which happens recursively on the stack for each
`ListEntry` in that list. This leads to a stack overflow.

This change adjusts that logic to use a new `filter` method. This is
somewhat less efficient, but not in a way that should matter in normal
workloads with small numbers of blocked puts and takes. The benefit is
far better correctness, and it totally avoids this bad access pattern.

The new `Queue.StressTest` method covers this by artificially creating
the bad access pattern. On some platforms, doing repeated takes that are
raced with 0 timeout creates the situation _quickly_. The stack traces
captured when this fails matches stack traces seen from core dumps in
the field.