BREAKING: Rework Queue, remove List
Queue is pretty overcomplicated for what it does. This can
make troubleshooting issues with it a bit difficult. This change
reworks `Queue`  to just use a `std::mutex` and structures from
the STL. This results in a simpler implementation.

Since the mutex is only used to guard the data structures themselves
and not used to pend the queue - it really has the same semantics as
before. So long as there isn't contention for access to
the queue's data structures a `std::mutex` will typically just
"lock" by setting an atomic flag. That's the same behavior
as before, essentially.

While under contention to the data structures themselves, this
queue _will_ block but that was true before as well. Previously
it would just block on an atomic spin lock. Typically a
`std::mutex` will, at this point, end up calling into the
kernel to do the blocking. That's probably the better and more
efficient thing to do anyway.

Since `MVar` is really just a special case of `Queue` we can also
rework the implementation there by making it use a `Queue`
internally instead.

This also removes the need for a `List` implementation in this
code base. There are other - better - implementations in the world
of a list with the semantics of structural sharing and atomic head
pushes and pops. Those are probably better conceived than the one
in this library anyway.

The `Queue` and `MVar` APIs remain unchanged. Since `List` was
removed though, this is a breaking API change.