feat: allow timers to be stolen
This change allows timers that have expired but having executed
yet to be stolen by idle schedulers.

Previously timers were "parked" to a scheduler and did not
participate in work-stealing. This meant that it was possible
for a single scheduler to get innundated with these timers which
would subsequently take priority over "normal" work because of
how the scheduler threads prioritize working through them.

With this change - in such a case the timers should start to
balance across all available schedulers because a scheduler
backed up by "normal work" _or_ expired timers should have any
timers that have expired get stolen. Since timers are always one-shot
any tasks that would subsequently delay on a timer again would
then be balanced to the new scheduler thread which will naturally
balance them out over time.

Part of this refactor is unifying how a scheduler thread fetches
work for itself with how the stealing works by making them
_both_ call a `steal_unsafe` method to implement this behavior.

This should functionally appear no different to observers, but
marking as a feature since its definitely a big upgrade to how
the scheduler behaves.