fix: never randomly select a 0ms idle time
When a thread is putting itself into idle it randomly selects
a wait time. After that wait time it'll resume and most likely
attempt to work steal from another thread.

Previously the scheduler could randomly select a time < 10ms with
even 0ms (so wake immediately) being a valid selection. To reduce
the contention from stealing - it's better to force some base amount
of idle time. I've chosen 10ms to be the base time which is still
a pretty quick wakeup time (100hz) and have retained the 100ms
ceiling time (10hz).