fix: avoid stealing tiny amounts of work
One of the benefits of work stealing is that work stays local
to a thread unless it really needs to spread out. Unfortunately
we had no guard in-place when the steal happened to prevent
tiny amounts of work to be continuously stolen.

This change adds a peek at the amount of available work and
uses a simple heuristic (`available > batch_size/2`) to detect
if there is enough work to justify the steal. This avoids a
case where a system with plenty of compute but not a lot of work
continuously shuffles around a tiny amount of work between
schedulers.