fix: always steal if a time is expired
When stealing, it's important to prioritize dealing with any
expired timers so that the time between when they expire and
when they are fire is a small as possible. In CPU-bound
scenarios, it's possible that a scheduler is processing work
when a timer expires and won't get around to firing it soon.

So, this change adjust the stealing heuristic to _always_ steal
if at least one timer is in the timer list which has expired
and is ready to fire.