BREAKING: Improve performance of the work stealing scheduler. (#99)
This change makes some important changes to the work stealing scheduler
motivated by performance issues discovered in a real-time multithreaded
workload. In such a case, the work stealing implemented would not
"spread out" across cores fast enough resulting in applications being
severely limited by single-threaded performance.

This was due to how each work thread would take an entire "batch" of
work (128 items by default) and not let those items be stolen - and
instead allow stealing only from a queue in front of that batch. In
practice, this meant a worker was allowed to claim ownership of up to
128 fibers at a time and never allow them to be stolen. This means that
work would not be spread out across cores - leading to a single worker
being overloaded while other works are searching for work.

This change makes some important changes to fundamentally resolve this
issue with a better implementation.

1. The batching behavior was removed. Workers take a single item from
their queue at a time to work, and other items
   can always be stolen.
2. A global "overflow" queue was introduced and the size of the queue
was limited. This allows overloaded workers to naturally push work back
into the overflow queue so that it can be easily be picked up by other
workers.
3. Stealing was adjusted to steal from both peers and the global queue.
Under normal circumstances stealing from peers is preferred, but with a
random chance workers will instead prefer the global queue to ensure it
is serviced as well.
4. Stealing was adjusted to take a half of the victim's queue at a time
so that work is better distributed between the stealer and the victim.
5. A new `ReadyQueue` implementation simplifies the locking of that
queue allowing it to be passed around as a thread-safe concurrent queue.
6. Threads searching for work now iterate through all peers rather than
looking at a single peer and then giving up.

These changes required some breaking API changes - but those changes are
unlikely to affect most users since they are mostly related to
parameters and methods on `SingleThreadScheduler` that clients were
unlikely to use (though they were able).

Since this change was breaking, it was a good time to do another minor
cleanup. The `batch_size` build time parameter
has been rename `cede_iterations` since its true purpose is to configure
`FiberImpl` to _force_ a cede even for continuations that can be
synchronously evaluated every so often. This new name better clarifies
its intent. I've also moved those items to a new `cask::config`
namespace to better clarify their role and make these constants easier
to spot in code.